// assets/core/boot/defines/boot.enum.ts
export enum BootState {
    FrameworkInit = "FrameworkInit",
    HotUpdate = "HotUpdate",
    SDKInit = "SDKInit",
    TokenCheck = "TokenCheck",
    PlatformLogin = "PlatformLogin",
    VerifyToken = "VerifyToken",
    ServerConnect = "ServerConnect",
    PullPlayerData = "PullPlayerData",
    EnterGame = "EnterGame",
    BootError = "BootError",
}